+++ test/components/component_test.rb
@@ -78,7 +78,7 @@ class PrimerComponentTest < Minitest::Test
     [Primer::TimelineItemComponent, {}, proc { |component| component.body { "Foo" } }],
     [Primer::Tooltip, { label: "More" }],
     [Primer::Alpha::UnderlineNav, { label: "aria label" }, proc { |component| component.tab(selected: true) { "Foo" } }],
-    [Primer::Alpha::Tooltip, { type: :label, for_id: "some-button", text: "Foo" }],
+    [Primer::Alpha::Tooltip, { type: :label, for_id: "some-button", text: "Foo" }]
   ].freeze
 
   def test_registered_components